v0
Odoo module to add a “Open Cash Drawer” button in the Point of Sale (POS) interface. Here’s a brief overview of the files you’ve shared:

open_drawer.js: This JavaScript file defines a new component OpenCashDrawer that extends PosComponent. It listens for a click event on the component, and when clicked, it shows a confirmation popup. If confirmed, it sends a GET request to ‘https://localhost:8443/open_cash_drawer’.

drawer_button.xml: This XML file defines the template for the “Open Cash Drawer” button. The button is styled with red color and includes a money icon.